# Prefer HTTP/2 when available