--- a/lib/ably/rest/middleware/fail_if_unsupported_mime_type.rb
+++ b/lib/ably/rest/middleware/fail_if_unsupported_mime_type.rb
@@ -7,7 +7,10 @@ module Ably
       class FailIfUnsupportedMimeType < Faraday::Response::Middleware
         def on_complete(env)
           unless env.response_headers['Ably-Middleware-Parsed'] == true
-            unless  (500..599).include?(env.status)
+            # Ignore empty body with success status code for no body response
+            return if env.body.to_s.empty? && env.status == 204
+
+            unless (500..599).include?(env.status)
               raise Ably::Exceptions::InvalidResponseBody,
                     "Content Type #{env.response_headers['Content-Type']} is not supported by this client library"
             end