--- a/lib/ably/version.rb
+++ b/lib/ably/version.rb
@@ -1,5 +1,5 @@
 module Ably
-  VERSION = '1.1.3'
+  VERSION = '1.1.4.rc'
   PROTOCOL_VERSION = '1.1'
 
   # Allow a variant to be configured for all instances of this client library