--- a/lib/ably/version.rb
+++ b/lib/ably/version.rb
@@ -1,5 +1,5 @@
 module Ably
-  VERSION = '1.1.4.rc'
+  VERSION = '1.1.4'
   PROTOCOL_VERSION = '1.1'
 
   # Allow a variant to be configured for all instances of this client library