HTTP/2 provides significantly improved throughput over a single connection, which in turn, reduces latency for clients publishing a high rate of messages that would otherwise be split of many HTTP/1 connections (only two in-flight requests supported).

Initial test results show a greater than 10x improvement in latencies when using HTTP/2 for highly concurrent requests (100).  Looking in Wireshark, as expected, this is down to the fact that a single connection is shared and only one TLS handshake is needed.  With HTTP1, around 50 were needed.

Using this trivial script:

```ruby
$LOAD_PATH.unshift(File.join(Dir.pwd, 'lib'))
require 'ably'
require 'parallel'
require 'benchmark'

rest = Ably::Rest.new(key: '[YOUR-KEY]', log_level: :debug)

CONCURRENCY=100
puts Benchmark.measure {
  Parallel.each(CONCURRENCY.times.to_a, in_threads: CONCURRENCY) { puts "Sending request"; puts "Response: #{rest.time}" }
}
```

### Before HTTP/2

First attempt examples (overlapping connection establishment attempts)

```
Response: 2020-03-22 20:25:44 +0000
  2.110579   0.152436   2.263015 (  2.775102)

Response: 2020-03-22 20:28:43 +0000
  2.122079   0.160033   2.282112 (  6.244351)
```

Second attempt examples (connection should exist)

```
Response: 2020-03-22 20:27:10 +0000
  2.281712   0.149430   2.431142 (  3.534759)

Response: 2020-03-22 20:27:22 +0000
  2.433119   0.217437   2.650556 (  2.908990)
```

### Using HTTP/2

First attempt example (overlapping connection establishment attempts)

```
Response: 2020-03-22 20:23:51 +0000
  1.688666   0.169178   1.857844 (  1.767157)
```

Second attempt examples (connection exists)

```
Response: 2020-03-22 20:24:25 +0000
  0.084862   0.032209   0.117071 (  0.244550)

Response: 2020-03-22 20:25:27 +0000
  0.103958   0.030839   0.134797 (  0.242053)
```

This fixes https://github.com/ably/ably-ruby/issues/192
Relates to https://github.com/ably/wiki/issues/309

Available as a RC candidate 1.1.4.rc - https://rubygems.org/gems/ably/versions/1.1.4.rc