@QuintinWillison 2020-03-23
Looks simple enough. I'm surprised at how little work was required to get that in place, but perhaps that's credit to previous architectural decisions (re middleware, etc..). Nice! 😀