@paddybyers 2020-04-01
So is the next step here to specify the client option, and make this behaviour conditional on that?